requirements.txt
# Core
openai>=1.40
python-dotenv>=1.0
tqdm>=4.66

# Automated checks
sentence-transformers>=2.7.0
scipy>=1.13.0
numpy>=1.26.0
pandas>=2.2.2
textstat>=0.7.4

# Grammar checking
language-tool-python>=2.7.0
